Updated story title rule
