+++ src/rules/util.js
@@ -162,35 +162,6 @@ module.exports.getSuperclassName = (classExpressionNode) => {
 };
 
 /**
- * Returns the expected story title for web (based on Instawork conventions) given the story filepath
- * e.g. web_frontend/<app_group>/booking-flow/components/tpb-card => Apps/BookingFlow/Components/TpbCard
- * e.g. web_frontend/common/src/components/iw-badge => Common/Components/IwBadge
- * @param {*} filePath
- * @returns expected story title
- */
-module.exports.getExpectedWebStoryTitle = (filePath) => {
-  const normalizedPath = filePath.replace(/\\/g, '/'); // Convert backslashes to forward slashes
-  const pathWithoutFile = normalizedPath.substring(0, normalizedPath.lastIndexOf('/'));
-  const pathParts = pathWithoutFile.split('/').filter((part) => part !== 'src'); // Split for parts and exclude "src"
-
-  // Extract the app group and app name
-  const webFrontendIndex = pathParts.indexOf('web_frontend');
-  const appGroup = pathParts[webFrontendIndex + 1];
-  const appName = pathParts[webFrontendIndex + 2];
-
-  const titlePrefix = appGroup === 'common' ? 'Common' : 'Apps';
-  let title = `${titlePrefix}/${humps.pascalize(appName)}`;
-  const remainingSections = pathParts.slice(webFrontendIndex + 3);
-
-  // Append the remaining sections to the title
-  remainingSections.forEach((section) => {
-    title += `/${humps.pascalize(section)}`;
-  });
-
-  return title;
-};
-
-/**
  * Returns whether or not a path points to a first-level common view index file
  * The pattern matches paths like "common/views/view-name/index.ts"
  * and rejects deeper paths like "common/views/view-name/subfolder/index.ts"
